Toutes les corrections coupe d'IDF

+++ src/Common/Action.Driver/ASensorsDriver.hpp
@@ -10,12 +10,12 @@ class RobotPos
 {
 public:
     int nbDetectedBots; //nb de position de robot detectée (dupliqué dans chaque)
-    int x; // position x de l'ADV en millimetres
-    int y; // position y de l'ADV en millimetres
+    float x; // position x de l'ADV en millimetres
+    float y; // position y de l'ADV en millimetres
     float theta_deg; //degre // angle de l'ADV par rapport à l'avant du robot (devant axe x = 0 degré, gauche (axe y) = >0deg, droite (axe y) = <0deg)
-    int d; //dist en mm à partir du centre du robot jusquau centre du robot ADV
+    float d; //dist en mm à partir du centre du robot jusquau centre du robot ADV
 
-    RobotPos(int nb, int x_, int y_, float a_, int d_)
+    RobotPos(int nb, float x_, float y_, float a_, float d_)
     {
         nbDetectedBots = nb;
         x = x_;
