+++ src/Driver-SIMU/SensorsDriver.hpp
@@ -46,7 +46,7 @@ public:
     int sync(); //synchronise les données avec la balise
     ASensorsDriver::bot_positions getvPositionsAdv(); //retourne les dernieres positions connues
 
-    RobotPos transformPosTableToPosRobot(int nb, int x_table, int y_table);
+    RobotPos transformPosTableToPosRobot(int nb, float x_table, float y_table);
 
 
     int rightSide();
